Remove redundant fakecaclient.
