+++ security/pkg/caclient/protocol/protocol.go
@@ -23,7 +23,7 @@ import (
 	"google.golang.org/grpc"
 
 	//"istio.io/istio/pkg/log"
-	"istio.io/istio/security/pkg/platform"
+
 	pb "istio.io/istio/security/proto"
 )
 
@@ -31,12 +31,6 @@ import (
 type CAProtocol interface {
 	// SendCSR send CSR request to the CA server.
 	SendCSR(*pb.CsrRequest) (*pb.CsrResponse, error)
-}
-
-// CAGrpcClient is for implementing the GRPC client to talk to CA.
-type CAGrpcClient interface {
-	// Send CSR to the CA and gets the response or error.
-	SendCSR(*pb.CsrRequest, platform.Client, string) (*pb.CsrResponse, error)
 }
 
 // GrpcConnection implements CAProtocol talking to CA via gRPC.
@@ -69,40 +63,4 @@ func (c *GrpcConnection) SendCSR(req *pb.CsrRequest) (*pb.CsrResponse, error) {
 // Close closes the gRPC connection.
 func (c *GrpcConnection) Close() error {
 	return c.connection.Close()
-}
-
-// FakeProtocol is a fake for testing, implements CAProtocol interface.
-type FakeProtocol struct {
-	counter int
-	resp    *pb.CsrResponse
-	errMsg  string
-}
-
-// NewFakeProtocol returns a FakeProtocol with configured response and expected error.
-func NewFakeProtocol(response *pb.CsrResponse, err string) *FakeProtocol {
-	return &FakeProtocol{
-		resp:   response,
-		errMsg: err,
-	}
-}
-
-// SendCSR returns the result based on the predetermined config.
-func (f *FakeProtocol) SendCSR(req *pb.CsrRequest) (*pb.CsrResponse, error) {
-	f.counter++
-	if f.counter > 8 {
-		return nil, fmt.Errorf("terminating the test with errors")
-	}
-
-	if f.errMsg != "" {
-		return nil, fmt.Errorf(f.errMsg)
-	}
-	if f.resp == nil {
-		return &pb.CsrResponse{}, nil
-	}
-	return f.resp, nil
-}
-
-// InvokeTimes returns the times that SendCSR has been invoked.
-func (f *FakeProtocol) InvokeTimes() int {
-	return f.counter
 }
