+++ security/pkg/nodeagent/vm/nodeagent_test.go
@@ -22,7 +22,7 @@ import (
 
 	"istio.io/istio/pkg/log"
 	"istio.io/istio/security/pkg/caclient"
-	"istio.io/istio/security/pkg/caclient/protocol"
+	"istio.io/istio/security/pkg/caclient/protocol/mock"
 	"istio.io/istio/security/pkg/platform"
 	mockpc "istio.io/istio/security/pkg/platform/mock"
 	"istio.io/istio/security/pkg/util"
@@ -56,7 +56,7 @@ func TestStartWithArgs(t *testing.T) {
 	testCases := map[string]struct {
 		config      *Config
 		pc          platform.Client
-		caProtocol  *protocol.FakeProtocol
+		caProtocol  *mock.FakeProtocol
 		certUtil    util.CertUtil
 		expectedErr string
 		sendTimes   int
@@ -65,7 +65,7 @@ func TestStartWithArgs(t *testing.T) {
 		"Success": {
 			config:      &generalConfig,
 			pc:          mockpc.FakeClient{nil, "", "service1", "", []byte{}, "", true},
-			caProtocol:  protocol.NewFakeProtocol(&pb.CsrResponse{IsApproved: true, SignedCert: signedCert, CertChain: certChain}, ""),
+			caProtocol:  mock.NewFakeProtocol(&pb.CsrResponse{IsApproved: true, SignedCert: signedCert, CertChain: certChain}, ""),
 			certUtil:    mockutil.FakeCertUtil{time.Duration(0), nil},
 			expectedErr: "node agent can't get the CSR approved from Istio CA after max number of retries (3)",
 			sendTimes:   12,
@@ -73,14 +73,14 @@ func TestStartWithArgs(t *testing.T) {
 		},
 		"Config Nil error": {
 			pc:          mockpc.FakeClient{nil, "", "service1", "", []byte{}, "", true},
-			caProtocol:  protocol.NewFakeProtocol(nil, ""),
+			caProtocol:  mock.NewFakeProtocol(nil, ""),
 			expectedErr: "node Agent configuration is nil",
 			sendTimes:   0,
 		},
 		"Platform error": {
 			config:      &generalConfig,
 			pc:          mockpc.FakeClient{nil, "", "service1", "", []byte{}, "", false},
-			caProtocol:  protocol.NewFakeProtocol(nil, ""),
+			caProtocol:  mock.NewFakeProtocol(nil, ""),
 			expectedErr: "node Agent is not running on the right platform",
 			sendTimes:   0,
 		},
@@ -103,42 +103,42 @@ func TestStartWithArgs(t *testing.T) {
 				LoggingOptions: log.DefaultOptions(),
 			},
 			pc:          mockpc.FakeClient{nil, "", "service1", "", []byte{}, "", true},
-			caProtocol:  protocol.NewFakeProtocol(nil, ""),
+			caProtocol:  mock.NewFakeProtocol(nil, ""),
 			expectedErr: "CSR creation failed (crypto/rsa: message too long for RSA public key size)",
 			sendTimes:   0,
 		},
 		"Getting agent credential error": {
 			config:      &generalConfig,
 			pc:          mockpc.FakeClient{nil, "", "service1", "", nil, "Err1", true},
-			caProtocol:  protocol.NewFakeProtocol(nil, ""),
+			caProtocol:  mock.NewFakeProtocol(nil, ""),
 			expectedErr: "request creation fails on getting agent credential (Err1)",
 			sendTimes:   0,
 		},
 		"SendCSR empty response error": {
 			config:      &generalConfig,
 			pc:          mockpc.FakeClient{nil, "", "service1", "", []byte{}, "", true},
-			caProtocol:  protocol.NewFakeProtocol(nil, ""),
+			caProtocol:  mock.NewFakeProtocol(nil, ""),
 			expectedErr: "node agent can't get the CSR approved from Istio CA after max number of retries (3)",
 			sendTimes:   4,
 		},
 		"SendCSR returns error": {
 			config:      &generalConfig,
 			pc:          mockpc.FakeClient{nil, "", "service1", "", []byte{}, "", true},
-			caProtocol:  protocol.NewFakeProtocol(nil, "error returned from CA"),
+			caProtocol:  mock.NewFakeProtocol(nil, "error returned from CA"),
 			expectedErr: "node agent can't get the CSR approved from Istio CA after max number of retries (3)",
 			sendTimes:   4,
 		},
 		"SendCSR not approved": {
 			config:      &generalConfig,
 			pc:          mockpc.FakeClient{nil, "", "service1", "", []byte{}, "", true},
-			caProtocol:  protocol.NewFakeProtocol(&pb.CsrResponse{IsApproved: false}, ""),
+			caProtocol:  mock.NewFakeProtocol(&pb.CsrResponse{IsApproved: false}, ""),
 			expectedErr: "node agent can't get the CSR approved from Istio CA after max number of retries (3)",
 			sendTimes:   4,
 		},
 		"SendCSR parsing error": {
 			config: &generalConfig,
 			pc:     mockpc.FakeClient{nil, "", "service1", "", []byte{}, "", true},
-			caProtocol: protocol.NewFakeProtocol(&pb.CsrResponse{
+			caProtocol: mock.NewFakeProtocol(&pb.CsrResponse{
 				IsApproved: true, SignedCert: signedCert, CertChain: []byte{},
 			}, ""),
 			certUtil:    mockutil.FakeCertUtil{time.Duration(0), fmt.Errorf("cert parsing error")},
