+++ security/pkg/probe/controller_test.go
@@ -25,6 +25,7 @@ import (
 
 	"istio.io/istio/pkg/probe"
 	"istio.io/istio/security/pkg/caclient/protocol"
+	"istio.io/istio/security/pkg/caclient/protocol/mock"
 	"istio.io/istio/security/pkg/pki/ca"
 	"istio.io/istio/security/pkg/pki/util"
 	pb "istio.io/istio/security/proto"
@@ -73,7 +74,7 @@ func TestGcpGetServiceIdentity(t *testing.T) {
 
 	for id, c := range testCases {
 		fakeProvider := func(_ string, _ []grpc.DialOption) (protocol.CAProtocol, error) {
-			return protocol.NewFakeProtocol(c.resp, c.err), nil
+			return mock.NewFakeProtocol(c.resp, c.err), nil
 		}
 		// test liveness probe check controller
 		controller, err := NewLivenessCheckController(
